+++ extensions/go-to-retrace-timestamp/CHANGELOG.md
@@ -1,4 +1,8 @@
-# Go to Retrace Timestamp Changelog
+# Retrace Quick Actions Changelog
+
+## [Unreleased]
+- Rename extension display title to `Retrace Quick Actions`
+- Add `Search Retrace` command that opens `retrace://search?q=<query>`
 
 ## [Initial Version] - 2026-02-09
 - Add `Go to Retrace Timestamp` command with natural language parsing